fix(time): 分境界スケジューラを shared/time に集約する
intent(minute-clock): 同じ壁時計を見せる面が 2 つあり、分境界に同期する仕組みを
それぞれが持っていた。スケジューラはどちらの feature のアイデンティティでもない
decision(minute-clock): 表示面は統合せず、スケジューラだけを shared/time へ置く。
SidebarClock が答えるのは「今何時か」、進捗バーが答えるのは「今日がどこまで進んだか」で
担う問いが違う。共有するのは仕組みであって機能ではない
decision(sidebar-clock): デジタル表示は Intl ではなく Temporal の ISO 出力にする。
Intl.DateTimeFormat は Date 用と Temporal 型用で別の format record を持ち
([[DateTimeFormat]] と [[TemporalPlainTimeFormat]])、後者は生成時に locale の
available pattern から選び直される。ja-JP の time pattern は numeric hour (H:mm) なので
hour:"2-digit" の要求はそこで落ち、09:36 が 9:36、00:05 が 0:05 になる
(resolvedOptions().hour は [[DateTimeFormat]] 側を映すため 2-digit のまま)。
オプションで変えられる段階ではないため、区切り文字の locale 追従
(fi-FI / da-DK / id-ID の "." が ":" になる) より桁の固定を採る
decision(shallow-ref): 値は shallowRef で持つ。Temporal の各型は内部スロットを持つ
exotic object で、Vue が proxy 化の対象を Object / Array / Map / Set 系に限っているため
実際には包まれないが、ref のままだと実装詳細への暗黙の依存になる
learned(use-timeout-fn): 張り直しに stop() は要らない。VueUse 14.4.0 の useTimeoutFn は
start() の先頭で clear() を呼び、setTimeout のコールバック内で timer を undefined に
戻して isPending も false にしている
constraint(day-progress): バー幅の現在値は max-w-md + px-40。日付ラベル 2 個と gap で
約 130px を使うため、帯の実効幅を 320px 前後に保つには外枠を広げる必要がある
