+++ apps/renderer/src/features/layout/TitleBar.vue
@@ -17,7 +17,7 @@
   視覚中心がずれ、ドラッグでウィンドウを動かすたびに目線が揺れる
 - 中央に出すのは repo / worktree 名ではなく 1 日の進捗バー。repo と branch は
   サイドバーと各ペインが常時見せているため帯に重ねて出す価値が薄く、代わりに
-  どこからも読めない情報（時刻）を置く。OS のウィンドウ一覧向けのネイティブ
+  1 日のどこにいるかという別種の情報を置く。OS のウィンドウ一覧向けのネイティブ
   タイトルは別経路（`useTitleContextSync`）が repo / branch を出し続ける
 </doc>
