+++ apps/renderer/src/features/layout/dayProgress.ts
@@ -5,8 +5,11 @@
  * 位置は「時刻 → パーセント」の線形写像だけで決まる。
  *
  * このモジュールは時刻の値を数値でしか受け取らない。現在時刻の取得や日付の
- * 整形は `Temporal` を使う呼び出し側が持つ。テストランナー（bun 1.3.14）は
- * まだ `Temporal` を持たないため、境界をここに引いて座標計算を検証可能に保つ。
+ * 整形は `Temporal` を使う呼び出し側が持つ。テストランナー（bun）が `Temporal` を
+ * 持たないため、境界をここに引いて座標計算だけでも検証可能に保つ。
+ *
+ * TODO: bun が `Temporal` を持つ版になったらこの分割は不要になり、呼び出し側の日付演算にも
+ * テストを書ける。
  */
 
 /** バーが表す 1 日の長さ（時間）。バー幅 100% がこの時間に対応する。 */
