+++ apps/renderer/src/shared/time/index.ts
@@ -1,4 +1,5 @@
 export { formatRelativeAge, isoToUnixSec } from "./relativeAge";
+export { useMinuteClock } from "./useMinuteClock";
 export {
   formatAbsoluteTime,
   formatCompactTime,
